Add crystal to USB
This is a real headache to find a crystal that matches the chip's
specs (50 ohm ESR) and the backplane 0.125" height restriction.
We're pushing a bit with this crystal but from my reading of TI's
supplemental guide for crystal selection for their high-speed USB
chip ( http://www.ti.com/lit/an/slla122/slla122.pdf ) it should
work.