Arduino: WIP, I should commit more frequently :(
The primary change in this commit is to extend the 3.3V/5V selection to the whole
board. The nice observation is that the two MCUs are 2.7-5.5 and 1.8-5.5 VCC tolerant,
so can do run-time selection of core voltage with a little bit of careful design. This
lets the seinduino look either like a Due (3.3V) or a Uno (5V).

The next mission is to replace that reset trigger circuit with a one-shot so that we
aren't cooking 5mA all the time.