added a read ready function.
I'm not convinced this is the right way to do this, but it works.
Going forward we should either 1) remove the need for read_ready()
entirely or 2) fix the underlying i2c code so that you can clock
stretch on a read call, call
read_ready() from the callback, then finish the read.
I think Brad intended number 2 originally.