debug_backplane: add debug USB hub **sch only**
attn @adkinsjd: @brghena had the good idea to add a second FTDI for
the debug radio, giving one serial connection for the "radio uplink"
and one for "radio debug", where the former handles all the data
tunneling and the latter is for debugging the fake radio itself.
I've added that because the marginal cost of another FTDI chip is
minimal and we still have another open port on the USB hub if we
want it for something, but wanted to bring that change to your
attention.

TODO: Deal with the power-on-reset and possible runtime reset for
the USB hub.