Changed isolator voltage source
Left side of the isolator is now connected to VCC_CONTROLLER, which is
the source of _almost_ all of its I/O lines. The except is the indicator
LED, which is powered from VCC_LINUX_3V3. In that case though, there's a
diode preventing current from flowing from Controller to Linux domain if
Linux is disabled.